including all tensorflow adds.

+++ gameengine.py
@@ -10,8 +10,8 @@ class GameEngine:
 
     def getBoardFreeSquares(self, snake):
         free_squares = []
-        for i in range(0, config.DEFAULT_WINDOW_WIDTH/config.STEP_SIZE):
-            for j in range(0, config.DEFAULT_WINDOW_HEIGHT/config.STEP_SIZE):
+        for i in range(0, int(config.DEFAULT_WINDOW_WIDTH/config.STEP_SIZE)):
+            for j in range(0, int(config.DEFAULT_WINDOW_HEIGHT/config.STEP_SIZE)):
                 if not (config.STEP_SIZE*i in snake.x and config.STEP_SIZE*j in snake.y):
                     free_squares.append((config.STEP_SIZE*i,config.STEP_SIZE*j))
         return free_squares
